Refactor and move controls to above table
