+++ src/rubintv/static/stylesheets/calendar.sass
@@ -1,5 +1,6 @@
 .calendar
-  flex-basis: 50%
+  width: 50%
+  min-width: 600px
   padding-top: 4em
   margin-top: -2em
