+++ src/rubintv/static/stylesheets/main.sass
@@ -156,9 +156,16 @@ h3
 
 .channel-grid-heading
 	background: #fff
-	padding-bottom: 6em
+	padding-top: 1em
 	position: sticky
 	top: 0
+	display: flex
+	flex-direction: row
+	align-items: center
+	height: 3.7em
+	z-index: 23
+	// h3
+	// 	margin-bottom: 0
 
 .per-day-channels
 	margin-bottom: 1em
@@ -173,7 +180,7 @@ h3
 .grid-title
 	text-align: left
 	position: sticky
-	top: 6em
+	top: 11em
 	&.sideways
 		max-width: 40px
 		// transform: rotate(270deg)
@@ -191,18 +198,42 @@ h3
 		flex-direction: column
 		justify-content: center
 
+.table-spacer
+	background: #fff
+	height: 8em
+	position: sticky
+	top: 4.4em
 
 .table-panel
 	background: #eee
 	padding: 1.2em 19px
 	border-radius: 20px
+	margin-left: 1em
 
-.table-control-title::after
-	content: "\25bc"
-	margin-left: 4px
+.table-control-button::after
+	content: "∆"
+	margin-left: 1em
+	transform-origin: 50% 50%
+	transform: rotate(90deg)
+	display: inline-block
+	width: 0.7em
+	transition: 0.3s
 
 .table-controls
-	margin-top: 1em
+	display: none
+	position: absolute
+	padding: 20px
+	margin: -3px -21px
+	border: 2px #fff solid
+	border-top: none
+	background: #eee
 
-.table-control-title
-	margin-bottom: 0
+.table-panel.open
+	.table-controls
+		display: block
+	.table-control-button::after
+		transform: rotate(180deg)
+
+.table-control-button
+	border: none
+	background: inherit
