+++ src/rubintv/templates/cameras/camera.jinja
@@ -47,5 +47,5 @@
 
 {% block footer_scripts %}
   <script src="//ajax.googleapis.com/ajax/libs/jquery/1.9.1/jquery.min.js"></script>
-  <script type=text/javascript src="/rubintv/static/js/refresh-daytable.js"></script>
+  <script type="module" src="/rubintv/static/js/refresh-daytable.js"></script>
 {% endblock footer_scripts  %}
